Initial version
Summary:
A static homepage and a solid system to expand it.

Build system is based on the Gulp build of ZURB Foundation template
https://github.com/zurb/foundation-zurb-template.

It could evolve to host documentation or OAuth login capability.
Or it could stay as is. Versatile use was a design choice.

Heading font has been chosen for the g and the ae ligature,
Kreon by Julia Petretta.

The background is a Beowulf manuscript at the British Library.

Test Plan:
* `npm start`
* `jsonlint`
* `npm run build`

Reviewers: dereckson

Differential Revision: https://devcentral.nasqueron.org/D536