Set call for actions buttons back to 100% in smaller viewports
Summary:
In a first design iteration, the button was at 100% width.

When switched from an element `a` to an element `input` to submit
a form to build the target URL in 6985ba335c86, we lost this feature,
welcome to be a more bold call for action.

This commit fixes a regression introduced in 6985ba335c86.

Test Plan: Tested with a view port < 640 px.

Reviewers: dereckson

Differential Revision: https://devcentral.nasqueron.org/D559